+++ blocklets/ai-studio/api/dev.ts
@@ -1,3 +1,4 @@
+import { projectCronManager } from '@api/libs/cron-jobs';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { setupClient } from 'vite-plugin-blocklet';
 
@@ -6,4 +7,11 @@ import { app, server } from './src';
 setupClient(app, {
   server,
   importMetaHot: import.meta.hot,
+});
+
+import.meta.hot?.on('vite:beforeFullReload', () => {
+  projectCronManager.destroy();
+});
+import.meta.hot?.dispose(() => {
+  projectCronManager.destroy();
 });