+++ blocklets/ai-studio/api/src/routes/project.ts
@@ -2,6 +2,7 @@ import fs from 'fs';
 import { mkdir, mkdtemp, readFile, rm } from 'fs/promises';
 import { basename, dirname, isAbsolute, join } from 'path';
 
+import { projectCronManager } from '@api/libs/cron-jobs';
 import { Config } from '@api/libs/env';
 import { NoPermissionError, NotFoundError } from '@api/libs/error';
 import { sampleIcon } from '@api/libs/icon';
@@ -10,6 +11,7 @@ import logger from '@api/libs/logger';
 import AgentInputSecret from '@api/store/models/agent-input-secret';
 import {
   ProjectSettings,
+  ResourceProject,
   fileToYjs,
   isAssistant,
   nextAssistantId,
@@ -17,6 +19,7 @@ import {
   variableToYjs,
 } from '@blocklet/ai-runtime/types';
 import { copyRecursive } from '@blocklet/ai-runtime/utils/fs';
+import { Map, getYjsValue } from '@blocklet/co-git/yjs';
 import { call } from '@blocklet/sdk/lib/component';
 import config from '@blocklet/sdk/lib/config';
 import { user } from '@blocklet/sdk/lib/middlewares';
@@ -30,10 +33,10 @@ import omit from 'lodash/omit';
 import omitBy from 'lodash/omitBy';
 import pick from 'lodash/pick';
 import uniqBy from 'lodash/uniqBy';
-import { joinURL, parseAuth, parseURL } from 'ufo';
+import { parseAuth, parseURL } from 'ufo';
 import { parse } from 'yaml';
 
-import { getProjectFromResource, getResourceProjects } from '../libs/resource';
+import { resourceManager } from '../libs/resource';
 import {
   ensureComponentCallOrPromptsAdmin,
   ensureComponentCallOrPromptsEditor,
@@ -42,15 +45,14 @@ import {
 import Project, { nextProjectId } from '../store/models/project';
 import {
   ASSETS_DIR,
-  CONFIG_FILENAME,
-  CONFIG_FILE_KEY,
-  CONFIG_FOLDER,
+  CONFIG_FILE_PATH,
+  COPY_REPO_FILES,
+  CRON_FILE_PATH,
   LOGO_FILENAME,
-  OLD_SETTINGS_FILE,
+  OLD_PROJECT_FILE_PATH,
+  PROJECT_FILE_PATH,
   ProjectRepo,
-  SETTINGS_FILE,
-  VARIABLE_FILENAME,
-  VARIABLE_KEY,
+  VARIABLE_FILE_PATH,
   autoSyncIfNeeded,
   defaultBranch,
   defaultRemote,
@@ -274,7 +276,7 @@ export function projectRoutes(router: Router) {
         const branches = await repository.listBranches();
 
         const working = await repository.working({ ref: project.gitDefaultBranch });
-        const projectSetting = working.syncedStore.files[SETTINGS_FILE] as ProjectSettings | undefined;
+        const projectSetting = working.syncedStore.files[PROJECT_FILE_PATH] as ProjectSettings | undefined;
 
         const users = await getAuthorsOfProject({
           projectId: project.id,
@@ -289,11 +291,11 @@ export function projectRoutes(router: Router) {
       })
     );
 
-    const resourceTemplates = (await getResourceProjects('template')).map((i) => ({
+    const resourceTemplates = (await resourceManager.getProjects({ type: 'template' })).map((i) => ({
       ...i.project,
       blockletDid: i.blocklet.did,
     }));
-    const resourceExamples = (await getResourceProjects('example')).map((i) => ({
+    const resourceExamples = (await resourceManager.getProjects({ type: 'example' })).map((i) => ({
       ...i.project,
       blockletDid: i.blocklet.did,
     }));
@@ -359,10 +361,11 @@ export function projectRoutes(router: Router) {
     const query = await logoQuerySchema.validateAsync(req.query, { stripUnknown: true });
 
     if (query.blockletDid) {
-      const resource = await getProjectFromResource({ blockletDid: query.blockletDid, projectId });
+      const resource = await resourceManager.getProject({ blockletDid: query.blockletDid, projectId });
 
-      if (resource?.gitLogoPath && (await exists(resource.gitLogoPath))) {
-        res.sendFile(resource.gitLogoPath);
+      const logoPath = resource?.dir && join(resource.dir, LOGO_FILENAME);
+      if (logoPath && (await exists(logoPath))) {
+        res.sendFile(logoPath);
         return;
       }
 
@@ -413,12 +416,12 @@ export function projectRoutes(router: Router) {
     const repo = await ProjectRepo.load({ projectId });
     if (query.working) {
       const w = await repo.working({ ref: query.projectRef || project.gitDefaultBranch || defaultBranch });
-      settings = w.syncedStore.files[SETTINGS_FILE] as ProjectSettings;
+      settings = w.syncedStore.files[PROJECT_FILE_PATH] as ProjectSettings;
     } else {
       try {
         const { blob } = await repo.readBlob({
           ref: project.gitDefaultBranch || defaultBranch,
-          filepath: SETTINGS_FILE,
+          filepath: PROJECT_FILE_PATH,
         });
         const str = Buffer.from(blob).toString();
         settings = parse(str);
@@ -528,7 +531,11 @@ export function projectRoutes(router: Router) {
     if (templateId) {
       // create project from resource blocklet
       if (blockletDid) {
-        const resource = await getProjectFromResource({ projectId: templateId, type: ['template', 'example'] });
+        const resource = await resourceManager.getProject({
+          blockletDid,
+          projectId: templateId,
+          type: ['template', 'example'],
+        });
 
         if (resource) {
           project = await createProjectFromTemplate(resource, {
@@ -559,7 +566,7 @@ export function projectRoutes(router: Router) {
         const template = projectTemplates.find((i) => i.project.id === templateId);
         if (template) {
           project = await createProjectFromTemplate(
-            { ...template, assistants: [] },
+            { ...template, agents: [] },
             { name, description, author: req.user! }
           );
         }
@@ -577,6 +584,8 @@ export function projectRoutes(router: Router) {
         description,
       });
     }
+
+    projectCronManager.reloadProjectJobs(project.id);
 
     res.json(project);
   });
@@ -616,8 +625,8 @@ export function projectRoutes(router: Router) {
             Buffer.from(
               (
                 await git
-                  .readBlob({ fs, gitdir, oid, filepath: SETTINGS_FILE })
-                  .catch(() => git.readBlob({ fs, gitdir, oid, filepath: OLD_SETTINGS_FILE }))
+                  .readBlob({ fs, gitdir, oid, filepath: PROJECT_FILE_PATH })
+                  .catch(() => git.readBlob({ fs, gitdir, oid, filepath: OLD_PROJECT_FILE_PATH }))
               ).blob
             ).toString()
           )
@@ -658,12 +667,12 @@ export function projectRoutes(router: Router) {
               await repository
                 .readBlob({
                   ref: originDefaultBranch!,
-                  filepath: SETTINGS_FILE,
+                  filepath: PROJECT_FILE_PATH,
                 })
                 .catch(() =>
                   repository.readBlob({
                     ref: originDefaultBranch!,
-                    filepath: OLD_SETTINGS_FILE,
+                    filepath: OLD_PROJECT_FILE_PATH,
                   })
                 )
             ).blob
@@ -721,7 +730,7 @@ export function projectRoutes(router: Router) {
     if (!isNil(name) || !isNil(description)) {
       const repository = await getRepository({ projectId });
       const working = await repository.working({ ref: project.gitDefaultBranch });
-      const projectSetting = working.syncedStore.files[SETTINGS_FILE] as ProjectSettings | undefined;
+      const projectSetting = working.syncedStore.files[PROJECT_FILE_PATH] as ProjectSettings | undefined;
       if (projectSetting) {
         if (!isNil(name)) projectSetting.name = name;
         if (!isNil(description)) projectSetting.description = description;
@@ -765,6 +774,10 @@ export function projectRoutes(router: Router) {
     checkProjectPermission({ req, project });
 
     await project.destroy();
+    projectCronManager.destroyProjectJobs(projectId);
+
+    const repository = await getRepository({ projectId });
+    await repository.destroy();
 
     const root = repositoryRoot(projectId);
     await Promise.all([
@@ -959,7 +972,7 @@ export function projectRoutes(router: Router) {
       const repo = await getRepository({ projectId });
 
       const working = await repo.working({ ref });
-      const settings = working.syncedStore.files[SETTINGS_FILE];
+      const settings = working.syncedStore.files[PROJECT_FILE_PATH];
 
       const agent = await getAssistantFromRepository({
         repository: repo,
@@ -1131,10 +1144,20 @@ async function copyProject({
   const repo = await getRepository({ projectId: project.id, author: { name: author.fullName, email: author.did } });
   const workingDir = join(dirname(repo.root), `${project.id}.cooperative/${defaultBranch}/working`);
   await copyRecursive(srcWorking.workingDir, workingDir);
+
   const working = await repo.working({ ref: defaultBranch });
-  working.syncedStore.files[SETTINGS_FILE] ??= {};
-  const settings = working.syncedStore.files[SETTINGS_FILE] as ProjectSettings;
-  Object.assign(settings, await projectSettingsSchema.validateAsync(project.dataValues));
+
+  for (const filepath of COPY_REPO_FILES) {
+    working.syncedStore.files[filepath] ??= {};
+    const file = working.syncedStore.files[filepath];
+
+    Object.assign(file, (getYjsValue(srcWorking.syncedStore.files[filepath]) as Map<any>).toJSON());
+
+    if (filepath === PROJECT_FILE_PATH) {
+      Object.assign(file, await projectSettingsSchema.validateAsync(project.dataValues));
+    }
+  }
+
   await working.commit({
     ref: defaultBranch,
     branch: defaultBranch,
@@ -1146,11 +1169,7 @@ async function copyProject({
 }
 
 async function createProjectFromTemplate(
-  template: Omit<(typeof projectTemplates)[number], 'project'> & {
-    project: Omit<(typeof projectTemplates)[number]['project'], 'createdAt' | 'updatedAt'>;
-    assetsDir?: string;
-    gitLogoPath?: string;
-  },
+  template: Partial<ResourceProject>,
   {
     name,
     description,
@@ -1166,10 +1185,10 @@ async function createProjectFromTemplate(
   const project = await Project.create({
     ...omit(template.project, 'name', 'files', 'createdAt', 'updatedAt', 'pinnedAt'),
     id: nextProjectId(),
-    duplicateFrom: withDuplicateFrom ? template.project.id : undefined,
+    duplicateFrom: withDuplicateFrom ? template.project?.id : undefined,
     createdBy: author.did,
     updatedBy: author.did,
-    gitDefaultBranch: template.project.gitDefaultBranch || defaultBranch,
+    gitDefaultBranch: defaultBranch,
     name,
     description,
   });
@@ -1181,7 +1200,7 @@ async function createProjectFromTemplate(
 
   const working = await repository.working({ ref: defaultBranch });
 
-  for (const { parent, ...file } of template.assistants) {
+  for (const { parent, ...file } of template.agents ?? []) {
     const id = file.id || nextAssistantId();
     const assistant = fileToYjs({ ...file, id });
 
@@ -1201,26 +1220,32 @@ async function createProjectFromTemplate(
     working.syncedStore.tree[id] = parent.concat(`${id}.yaml`).join('/');
   }
 
-  working.syncedStore.tree[VARIABLE_KEY] = joinURL(CONFIG_FOLDER, VARIABLE_FILENAME);
-  working.syncedStore.files[VARIABLE_KEY] = {
-    type: 'variables',
+  working.syncedStore.tree[VARIABLE_FILE_PATH] = VARIABLE_FILE_PATH;
+  working.syncedStore.files[VARIABLE_FILE_PATH] = {
     variables: (template.memory?.variables ?? []).map(variableToYjs),
   };
-  working.syncedStore.tree[CONFIG_FILE_KEY] = joinURL(CONFIG_FOLDER, CONFIG_FILENAME);
-  working.syncedStore.files[CONFIG_FILE_KEY] = template.config || {};
+
+  working.syncedStore.tree[CONFIG_FILE_PATH] = CONFIG_FILE_PATH;
+  working.syncedStore.files[CONFIG_FILE_PATH] = template.config || {};
+
+  working.syncedStore.tree[CRON_FILE_PATH] = CRON_FILE_PATH;
+  working.syncedStore.files[CRON_FILE_PATH] = template.cron || {};
 
   const assetsDir = join(working.workingDir, 'assets/');
   await mkdir(assetsDir, { recursive: true });
-  if (template.assetsDir && (await exists(template.assetsDir))) {
-    await copyRecursive(join(template.assetsDir, '/'), assetsDir);
+  const templateAssetsDir = template.dir && join(template.dir, 'assets/');
+  if (templateAssetsDir && (await exists(templateAssetsDir))) {
+    await copyRecursive(templateAssetsDir, assetsDir);
   }
+
+  const iconPath = template.dir && join(template.dir, LOGO_FILENAME);
 
   await repository.commitWorking({
     ref: defaultBranch,
     branch: defaultBranch,
     message: 'First Commit',
     author: { name: author.fullName, email: author.did },
-    icon: template?.gitLogoPath || (await sampleIcon()),
+    icon: iconPath && (await exists(iconPath)) ? iconPath : await sampleIcon(),
   });
 
   return project;