+++ blocklets/ai-studio/api/src/templates/projects.ts
@@ -1,25 +1,18 @@
-import { Assistant, ConfigFile, Variable } from '@blocklet/ai-runtime/types';
+import { ResourceProject } from '@blocklet/ai-runtime/types';
 
 import { wallet } from '../libs/auth';
-import Project from '../store/models/project';
 
-export const projectTemplates: {
-  project: Partial<Project['dataValues']>;
-  assistants: (Assistant & { parent: string[] })[];
-  config?: Partial<ConfigFile>;
-  memory?: { variables?: Variable[] };
-}[] = [
+export const projectTemplates: (Partial<ResourceProject> & Required<Pick<ResourceProject, 'project'>>)[] = [
   {
     project: {
       id: '363299428078977024',
       name: 'blank',
       description: 'Start a project from scratch.',
-      gitDefaultBranch: 'main',
       createdBy: wallet.address,
       updatedBy: wallet.address,
-      createdAt: new Date('2023-09-30T12:23:04.603Z'),
-      updatedAt: new Date('2023-09-30T12:23:04.603Z'),
+      createdAt: '2023-09-30T12:23:04.603Z',
+      updatedAt: '2023-09-30T12:23:04.603Z',
     },
-    assistants: [],
+    agents: [],
   },
 ];