+++ blocklets/ai-studio/src/components/file-editor/execute-block.tsx
@@ -5,13 +5,14 @@ import { useLocaleContext } from '@arcblock/ux/lib/Locale/context';
 import {
   AssistantYjs,
   ConfigFileYjs,
+  CronFileYjs,
   ExecuteBlock,
   ExecuteBlockYjs,
   FileTypeYjs,
+  MemoryFileYjs,
   ProjectSettings,
   Role,
   Tool,
-  VariablesYjs,
   isAssistant,
 } from '@blocklet/ai-runtime/types';
 import { Map, getYjsValue } from '@blocklet/co-git/yjs';
@@ -806,7 +807,9 @@ function ToolItemView({
 
 type Option = {
   id: NonNullable<ExecuteBlock['tools']>[number]['id'];
-  type: Exclude<FileTypeYjs, { $base64: string } | VariablesYjs | ProjectSettings | ConfigFileYjs>['type'] | string;
+  type:
+    | Exclude<FileTypeYjs, { $base64: string } | MemoryFileYjs | ProjectSettings | ConfigFileYjs | CronFileYjs>['type']
+    | string;
   name?: any;
   from?: NonNullable<ExecuteBlock['tools']>[number]['from'];
   fromText?: string;